backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
twilio==8.3.0
PyMySQL==1.0.3
mysqlclient==2.1.1
python-dateutil==2.8.2